添加 uswgi、docker-entrypoint.sh

+++ app.py
@@ -6,21 +6,23 @@ from resource import api
 from models import db
 from config import Config
 
+app = Flask(__name__)
 
-def create_app(config):
-    app = Flask(__name__)
+
+def init_app(config):
     app.config.from_object(config)
     config.init_app(app)
     db.init_app(app)
     api.init_app(app)
 
-    with app.app_context():
-        db.drop_all()
-        db.create_all()
+    if app.debug:
+        with app.app_context():
+            db.drop_all()
+            db.create_all()
 
-    return app
 
+init_app(Config)
 
 if __name__ == '__main__':
-    app = create_app(Config)
     app.run()
+
